Extract contracts
