+++ packages/kernel/source/index.ts
@@ -3,7 +3,6 @@ export * as Bootstrap from "./bootstrap/index.js";
 export * as Enums from "./enums/index.js";
 export * as Environment from "./environment.js";
 export * as Ipc from "./ipc/index.js";
-export * as IpcWorker from "./ipc-worker/index.js";
 export * as Providers from "./providers/index.js";
 export * as Services from "./services/index.js";
 export * as Support from "./support/index.js";
