--- packages/kernel/source/ipc-worker/index.ts
@@ -1,1 +1,0 @@
-export * from "./ipc-worker.js";
